+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## Unreleased
+## 2.1.1 - 2015-02-25
 ### Added
 - `PreferenceInjector.inject(Fragment target)` method. The `Context` can be inferred from the `Fragment`.
 - `PreferenceInjector.inject(Service target)` method. The `Context` can be inferred from the `Service`.